+++ towerdiscord/integration_test.go
@@ -75,12 +75,12 @@ func TestIntegration_NoFiles(t *testing.T) {
 	bot.SetHook(testHook{t: t, wg: wg})
 	tow.RegisterMessenger(bot)
 	//tow.NewEntry("test %d", 123).Context(tower.F{"foo": "bar", "struct": foo{}}).Notify(ctx)
-	origin := tow.Wrap(foo{FooMessage: "something > something < something & good"}).Message("this is foo error").Context(tower.F{
+	origin := tow.Wrap(foo{FooMessage: "something > something < something & Bad Request"}).Code(400).Message("this is bad request error").Context(tower.F{
 		"light": map[string]any{"year": 2021, "month": "january"},
 		"bar":   "baz",
 	}).Freeze()
 	wrapped := tow.WrapFreeze(origin, "wrapping error")
-	_ = tow.WrapFreeze(wrapped, "wrapping error").Notify(ctx)
+	_ = tow.Wrap(wrapped).Message("wrapping error").Context(tower.F{"wrapping": 123}).Notify(ctx)
 	err := bot.Wait(ctx)
 	if err != nil {
 		t.Error(err)